--- a/pinned_reqs/install.txt
+++ b/pinned_reqs/install.txt
@@ -4,4 +4,4 @@ pandas==1.0.0
 sortedcontainers==2.0.0
 scikit-learn==0.20.0
 scipy==1.4.0
-threadpoolctl==2.0.0
+threadpoolctl==3.0.0