--- a/setup.py
+++ b/setup.py
@@ -18,7 +18,7 @@ install_requires = [
     "sortedcontainers>=2.0.0",  # Primarily used in SlidingBoundariesArchive.
     "scikit-learn>=0.20.0",  # Primarily used in CVTArchive.
     "scipy>=1.4.0",  # Primarily used in CVTArchive.
-    "threadpoolctl>=2.0.0",
+    "threadpoolctl>=3.0.0",
     "semantic-version>=2.10"
 ]