## Description

<!-- Provide a brief description of the PR's purpose here. -->

Previously, we placed threadpoolctl in the optimizer in order to force OpenBLAS to run with only one thread in CMA-ES. However, this meant that all emitters would be affected by OpenBLAS, including ones defined by users. This commit moves threadpoolctl out of the optimizer and into CMA-ES, so that we only limit OpenBLAS in our code.

In #51, we mentioned that moving threadpoolctl to CMA-ES failed due to the overhead of creating the context manager. The new decorator from threadpoolctl seems to resolve this.

Here are some approximate times from running the sphere example (`python examples/sphere.py cma_me_imp --dim 20` and `python examples/sphere.py cma_me_imp --dim 100`) with the different ways of calling threadpoolctl. Note that when threadpoolctl is disabled, the multithreading in OpenBLAS causes race conditions which slow down the execution (see #51).

| Method |  20-dim sphere | 100-dim sphere |
| ---------- | ------ | ----- |
| Only in CMA-ES (this PR) | 55s | 94s |
| In Optimizer (previous) | 55s | 98s |
| No threadpoolctl | 60s | 458s |

## TODO

<!-- Notable points that this PR has either accomplished or will accomplish. -->

## Questions

<!-- Any concerns or points of confusion? -->

## Status

- [x] I have read the guidelines in [CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have tested my code by running `pytest`
- [x] I have linted my code with `pylint`
- [x] I have added a one-line description of my change to the changelog in `HISTORY.md`
- [x] This PR is ready to go